[FEAT/REFACTOR] 예약 및 결제 취소 리팩토링, 예약 변경 기능 추가

See merge request 23_HF141/studious-backend!81

+++ src/main/java/nerds/studiousTestProject/payment/service/PaymentService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import nerds.studiousTestProject.common.exception.BadRequestException;
 import nerds.studiousTestProject.common.exception.NotFoundException;
-import nerds.studiousTestProject.convenience.entity.ConvenienceRecord;
 import nerds.studiousTestProject.convenience.repository.ConvenienceRecordRepository;
+import nerds.studiousTestProject.member.entity.member.MemberRole;
 import nerds.studiousTestProject.payment.dto.callback.request.DepositCallbackRequest;
 import nerds.studiousTestProject.payment.dto.virtual.response.VirtualAccountInfoResponse;
 import nerds.studiousTestProject.payment.entity.PaymentStatus;
@@ -19,14 +19,13 @@ import nerds.studiousTestProject.payment.util.PaymentGenerator;
 import nerds.studiousTestProject.reservation.dto.detail.response.ReservationDetailResponse;
 import nerds.studiousTestProject.reservation.entity.ReservationRecord;
 import nerds.studiousTestProject.reservation.repository.ReservationRecordRepository;
-import nerds.studiousTestProject.room.entity.Room;
-import nerds.studiousTestProject.studycafe.entity.Studycafe;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import java.util.List;
 
+import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.MISMATCH_CANCEL_PRICE;
 import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.MISMATCH_ORDER_ID;
 import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.MISMATCH_PRICE;
-import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.NOT_FOUND_CONVENIENCE_RECORD;
 import static nerds.studiousTestProject.payment.entity.PaymentMethod.VIRTUAL_ACCOUNT;
 import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.INVALID_PAYMENT_SECRET;
 import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.MISMATCH_PAYMENT_METHOD;
@@ -53,86 +52,85 @@ public class PaymentService {
 
 
     @Transactional
-    public ReservationDetailResponse confirmSuccess(String orderId, String paymentKey, Integer amount) {
-        Payment payment = findByOrderId(orderId);
+    public ReservationDetailResponse confirmSuccess(final String orderId, final String paymentKey, final Integer amount) {
+        Payment payment = findByOrderIdWithReservationAndPlace(orderId);
         validConfirmRequest(orderId, amount, payment);
-        PaymentResponseFromToss responseFromToss = paymentGenerator.requestToToss(new ConfirmSuccessRequest(orderId, paymentKey, amount), CONFIRM_URI);
+        final PaymentResponseFromToss responseFromToss = paymentGenerator.requestToToss(new ConfirmSuccessRequest(orderId, paymentKey, amount), CONFIRM_URI);
         payment.complete(responseFromToss.toPayment());
         ReservationRecord reservationRecord = payment.getReservationRecord();
         reservationRecord.completePay();
-        return createReservationDetailResponse(payment, reservationRecord);
+        return ReservationDetailResponse.of(reservationRecord, payment);
     }
 
-    private void validConfirmRequest(String orderId, Integer amount, Payment payment){
+    private void validConfirmRequest(final String orderId, final Integer amount, final Payment payment){
         if(!amount.equals(payment.getPrice())) throw new BadRequestException(MISMATCH_PRICE);
         if(!orderId.equals(payment.getOrderId())) throw new BadRequestException(MISMATCH_ORDER_ID);
     }
 
-    private ReservationDetailResponse createReservationDetailResponse(Payment payment, ReservationRecord reservationRecord) {
-        Room room = reservationRecord.getRoom();
-        Studycafe studycafe = room.getStudycafe();
-        return ReservationDetailResponse.of(reservationRecord, studycafe, room, payment);
-    }
-
     @Transactional
-    public VirtualAccountInfoResponse virtualAccount(String orderId, String paymentKey, Integer amount) {
+    public VirtualAccountInfoResponse virtualAccount(final String orderId, final String paymentKey, final Integer amount) {
         Payment payment = findByOrderId(orderId);
         validConfirmRequest(orderId, amount, payment);
-        PaymentResponseFromToss responseFromToss = paymentGenerator.requestToToss(new ConfirmSuccessRequest(orderId, paymentKey, amount), CONFIRM_URI);
+        final PaymentResponseFromToss responseFromToss = paymentGenerator.requestToToss(new ConfirmSuccessRequest(orderId, paymentKey, amount), CONFIRM_URI);
         validPaymentMethod(responseFromToss);
-        payment.complete(responseFromToss.toVitualAccountPayment());
+        payment.complete(responseFromToss.toVitualAccountPayment()); //complete 말고 다른 작명이 좋을 수도 있을 듯
         log.info("success payment ! payment status is {} and method is {}", responseFromToss.getStatus(), responseFromToss.getMethod());
         return VirtualAccountInfoResponse.from(payment);
     }
 
-    private void validPaymentMethod(PaymentResponseFromToss responseFromToss) {
+    private void validPaymentMethod(final PaymentResponseFromToss responseFromToss) {
         if (!responseFromToss.getMethod().equals(VIRTUAL_ACCOUNT.getValue())) {
             throw new BadRequestException(MISMATCH_PAYMENT_METHOD);
         }
     }
 
     @Transactional
-    public ConfirmFailResponse confirmFail(String message, String orderId){
-        Payment payment = findByOrderId(orderId);
-        ReservationRecord reservationRecord = payment.getReservationRecord();
-        ConvenienceRecord convenienceRecord = findConvenienceRecordByReservationRecord(reservationRecord);
-        convenienceRecordRepository.delete(convenienceRecord);
-        reservationRecordRepository.delete(reservationRecord);
+    public ConfirmFailResponse confirmFail(final String message, final String orderId){
+        final Payment payment = findByOrderIdWithReservation(orderId);
+        final ReservationRecord reservationRecord = payment.getReservationRecord();
+        convenienceRecordRepository.findAllByPayment(payment).stream().forEach(convenience -> convenienceRecordRepository.delete(convenience));
         paymentRepository.delete(payment);
+        if (!paymentRepository.existsByReservationRecord(reservationRecord)) {
+            reservationRecordRepository.delete(reservationRecord);
+        }
         return ConfirmFailResponse.of(message);
     }
 
-    private ConvenienceRecord findConvenienceRecordByReservationRecord(ReservationRecord reservationRecord) {
-        return convenienceRecordRepository.findByReservationRecord(reservationRecord).orElseThrow(()-> new NotFoundException(NOT_FOUND_CONVENIENCE_RECORD));
-    }
-
     @Transactional
-    public void cancel(CancelRequest cancelRequest, Long reservationId){
-        ReservationRecord reservationRecord = findReservationById(reservationId);
-        Payment payment = findByReservationRecord(reservationRecord);
-        validPaymentMethod(cancelRequest, payment);
-        PaymentResponseFromToss responseFromToss = paymentGenerator.requestToToss(cancelRequest, String.format(CANCEL_URI, payment.getPaymentKey()));
-        payment.cancel(responseFromToss);
+    public void cancel(final CancelRequest cancelRequest, final Long reservationId){
+        final ReservationRecord reservationRecord = findReservationById(reservationId);
+        final List<Payment> payments = findPaymentsByReservationRecord(reservationRecord);
+        payments.stream().forEach(payment -> validPaymentMethod(cancelRequest, payment));
+        //취소 금액 검증
+        int totalCancelPrice = 0;
+        for (Payment payment : payments) {
+            final PaymentResponseFromToss responseFromToss = paymentGenerator.requestToToss(cancelRequest, String.format(CANCEL_URI, payment.getPaymentKey()));
+            payment.cancel(responseFromToss, MemberRole.USER);
+            totalCancelPrice += responseFromToss.getTotalAmount();
+        }
+        if (totalCancelPrice != payments.stream().mapToInt(Payment::getPrice).sum()) {
+            throw new BadRequestException(MISMATCH_CANCEL_PRICE);
+        }
         reservationRecord.canceled(); //결제 취소 상태로 변경
     }
 
-    private void validPaymentMethod(CancelRequest cancelRequest, Payment payment) {
+    private List<Payment> findPaymentsByReservationRecord(final ReservationRecord reservationRecord) {
+        return paymentRepository.findAllByReservationRecord(reservationRecord);
+    }
+
+    private void validPaymentMethod(final CancelRequest cancelRequest, final Payment payment) {
         if (payment.getMethod().equals(VIRTUAL_ACCOUNT.getValue())) {
             cancelRequest.getRefundReceiveAccount().validRefundVirtualAccountPay();
         }
     }
 
-    private Payment findByReservationRecord(ReservationRecord reservationRecord) {
-        return paymentRepository.findByReservationRecord(reservationRecord).orElseThrow(() -> new NotFoundException(NOT_FOUND_PAYMENT));
-    }
-
-    private ReservationRecord findReservationById(Long reservationId) {
+    private ReservationRecord findReservationById(final Long reservationId) {
         return reservationRecordRepository.findById(reservationId).orElseThrow(()-> new NotFoundException(NOT_FOUND_RESERVATION_RECORD));
     }
 
-    public void processDepositByStatus(DepositCallbackRequest depositCallbackRequest) {
-        Payment payment = findByOrderId(depositCallbackRequest.getOrderId());
-        String status = depositCallbackRequest.getStatus();
+    public void processDepositByStatus(final DepositCallbackRequest depositCallbackRequest) {
+        Payment payment = findByOrderIdWithReservation(depositCallbackRequest.getOrderId());
+        final String status = depositCallbackRequest.getStatus();
         ReservationRecord reservationRecord = payment.getReservationRecord();
         if(isDepositError(payment, status)){ // 입금 오류
             //입금 오류에 관한 알림 전송
@@ -149,18 +147,28 @@ public class PaymentService {
         payment.updateStatus(PaymentStatus.valueOf(status));
     }
 
-    private boolean isDepositError(Payment payment, String status) {
+    private boolean isDepositError(final Payment payment, final String status) {
         return status.equals(WAITING_FOR_DEPOSIT.name()) && payment.getStatus().equals(DONE);
     }
 
-    private void validPaymentSecret(DepositCallbackRequest depositCallbackRequest, Payment payment) {
+    private void validPaymentSecret(final DepositCallbackRequest depositCallbackRequest, final Payment payment) {
         if (!payment.getSecret().equals(depositCallbackRequest.getSecret())) {
             throw new BadRequestException(INVALID_PAYMENT_SECRET);
         }
     }
 
-    private Payment findByOrderId(String orderId) {
+    private Payment findByOrderId(final String orderId) {
         return paymentRepository.findByOrderId(orderId).orElseThrow(() -> new NotFoundException(NOT_FOUND_PAYMENT));
     }
+
+    private Payment findByOrderIdWithReservation(final String orderId) {
+        return paymentRepository.findByOrderIdWithReservation(orderId).orElseThrow(() -> new NotFoundException(NOT_FOUND_PAYMENT));
+    }
+
+    private Payment findByOrderIdWithReservationAndPlace(final String orderId) {
+        return paymentRepository.findByOrderIdWithReservationAndPlace(orderId).orElseThrow(() -> new NotFoundException(NOT_FOUND_PAYMENT));
+    }
+
+
 
 }
